requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
APScheduler==3.9.1
yfinance==0.2.18
alpha_vantage==2.3.1
python-dotenv==0.21.0
requests==2.28.2
pandas>=2.0.0
numpy>=1.24.3
scipy>=1.11.3
